requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# RAG & Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
unstructured>=0.15.0

# Vector Store & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4

# Memory & Storage
sqlalchemy>=2.0.0
# sqlite3 is built-in to Python, no need to install

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
jinja2>=3.1.3
aiofiles>=23.2.1

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
tenacity>=8.2.3
tiktoken>=0.5.2

# Observability
prometheus-client>=0.19.0
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Data Generation
faker>=22.0.0
reportlab>=4.0.7
